Added the discard when hand full rule, game save, split display in two classes (controller and display) and did a game log. The program compiles but the build needs extensive testing.
