+++ Build 2/ConsoleFormat.h
@@ -7,6 +7,7 @@
 
 using namespace std;
 
+class Session;
 class ConsoleFormat
 {
