+++ Build 2/EpidemicCard.h
@@ -2,6 +2,8 @@
 #include "Card.h"
 #include <iostream>
 #include <string>
+#include <boost/archive/text_iarchive.hpp>
+#include <boost/archive/text_oarchive.hpp>
 
 using namespace std;
 
@@ -9,6 +11,13 @@ class EpidemicCard :
 	public Card
 {
 private:
+	friend class boost::serialization::access;
+	template<class Archive>
+	inline void EpidemicCard::serialize(Archive & ar, const unsigned int version)
+	{
+		ar & description;
+	}
+
 	string description;
 	
 public:
